new ext doesnt require all files; followups temporarily do. code quality boost for new ext
