+++ src/app/api/generate/stream/route.js
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server"
 import { createClient } from "@/lib/supabase/server"
-import { generateChromeExtensionStream } from "@/lib/codegen/generate-extension"
+import { generateChromeExtensionStream } from "@/lib/codegen/generate-extension-stream"
 import { REQUEST_TYPES } from "@/lib/prompts/request-types"
 import { PLAN_LIMITS, DEFAULT_PLAN } from "@/lib/constants"
 import { isContextLimitError } from "@/lib/services/openai-responses"
